How many days has each customer visited us
How many days has each customer visited the restaurant? I used count function to count the number of days that the customer visited the restaurant, also I used the distinct function to select each day, if distinct is not used and a customer visited the restaurant twice a day, the count function will count a day twice.